fix: issue #42 rename component
