+++ src/App.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
-import './App.scss';
+import './app.scss';
 // import ReactApp from './components/molucules/ReactApp/ReactApp';
-import IdInfo from 'components/molucules/IdInfo/IdInfo';
-import CommonFrame from 'components/molucules/CommonFrame/CommonFrame';
+import IdInfo from 'components/molucules/id-info/id-info';
+import ComponentFrame from 'components/molucules/component-frame/component-frame';
 
 function App() {
   return (
     <div className="App">
       {/* <ReactApp /> */}
-      <CommonFrame viewComponent={<IdInfo />} componentName="idInfo" />
+      <ComponentFrame viewComponent={<IdInfo />} componentName="id-info" />
     </div>
   );
 }
