+++ src/components/molucules/component-frame/component-frame.scss
@@ -23,7 +23,7 @@
     padding: 20px;
     font-size: 100%;
     margin-bottom: 10px;
-    &.idInfo {
+    &.id-info {
       display: block;
     }
   }
