+++ src/components/molucules/id-info/id-info.scss
@@ -1,3 +1,3 @@
-.IdInfo {
+.id-info {
   font-weight: bold;
 }
